Keep track of cards that have been clicked and fix images not displaying
Remove the text at the end of the photo links to make it end with a png
so the image actually shows up.
Also created a function to handle the clicks on each card that takes the
id of the card and makes changes to its click status and changes the
score.